ai-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deep-translator==1.11.4
uvicorn[standard]
ultralytics
mediapipe
deepface
torch
torchvision
torchaudio
transformers
tf-keras
ffmpeg-python
fpdf2
python-multipart
sqlalchemy
opencv-python-headless
speechrecognition
fastapi==0.115.12
h11==0.9.0
httpx==0.13.3
idna==2.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
